% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_cut.R
\name{pretty_cut}
\alias{pretty_cut}
\title{Convert Numeric to Nice Factor}
\usage{
pretty_cut(x, buckets = 20, inf.left = T, inf.right = T)
}
\arguments{
\item{x}{a numeric vector which is to be converted to a factor by cutting}

\item{buckets}{approximate number of intervals into which x is to be cut}

\item{inf.left}{logical indicating if the leftmost interval should expand to include -Inf}

\item{inf.right}{logical indicating if the rightmost interval should expand to include Inf}
}
\value{
A \code{\link{factor}} is returned.
}
\description{
Creates an ordered factor from a numeric vector input with "nice" endpoints.
}
