% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emblem.R
\name{theme_emblem}
\alias{theme_emblem}
\title{ggplot color theme based on Willis Towers Watson's Emblem}
\usage{
theme_emblem(base_size = 10, base_family = "sans",
  base_line_size = 0.5, base_rect_size = 0.5)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}
}
\description{
A theme that approximates the style of \emph{Emblem}
}
\details{
\code{theme_emblem} implements the standard bluish-gray background and yellow
legend theme in \emph{Emblem}
}
\references{
\itemize{
\item \href{https://www.towerswatson.com/en-US/Services/Tools/emblem}{Emblem}
\item \url{https://www.willistowerswatson.com/en-US/about-us/overview}
}
}
\concept{themes emblem}
